Added file upload and pdf creation mechanism

+++ mod2/controllers/admin.js
@@ -1,6 +1,5 @@
 const Product = require('../models/product');
-const { request } = require('express');
-const product = require('../models/product');
+const fileHelper = require('../util/file'); 
 const { validationResult } = require('express-validator/check');
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product',
@@ -15,15 +14,33 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req,res,next)=>{
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
     const errors = validationResult(req);
+    if(!image){
+        return res.status(422).render('admin/edit-product',
+        {
+            pageTitle: 'Add Product',
+            path:'/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not image!',
+            validationErrors: []
+        });            
+    }
+    const imageUrl = image.path;
+
     if(!errors.isEmpty()){
         return res.status(422).render('admin/edit-product',
         {
             pageTitle: 'Add Product',
-            path:'/admin/edit-product',
+            path:'/admin/add-product',
             editing: false,
             hasError: true,
             product: {
@@ -51,7 +68,27 @@ exports.postAddProduct = (req,res,next)=>{
             console.log("Created Product");
             res.redirect('/admin/products');
         })
-        .catch((err) => console.log(err));
+        .catch((err) => {
+            // return res.status(500).render('admin/edit-product',
+            // {
+            //     pageTitle: 'Add Product',
+            //     path:'/admin/add-product',
+            //     editing: false,
+            //     hasError: true,
+            //     product: {
+            //         title: title,
+            //         imageUrl: imageUrl,
+            //         price: price,
+            //         description: description
+            //     },
+            //     errorMessage: 'Database operation failed, please try again',
+            //     validationErrors: []
+            // });            
+            // res.redirect('/500');
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+        });
 
 }
 
@@ -80,14 +117,19 @@ exports.getEditProduct = (req, res, next) => {
                 validationErrors: []
             });            
         })
-        .catch(err => console.log(err))
+        .catch(err => {
+            console.log(err)
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+        })
 
 }
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDesc = req.body.description;
     const errors = validationResult(req);
     if(!errors.isEmpty()){
@@ -99,7 +141,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDesc,
                 _id: prodId
@@ -116,13 +157,21 @@ exports.postEditProduct = (req, res, next) => {
         product.title = updatedTitle;
         product.price = updatedPrice;
         product.description = updatedDesc;
-        product.imageUrl = updatedImageUrl;
+        if(image){
+            fileHelper.deleteFile(product.imageUrl);
+            product.imageUrl = image.path;
+        }
         return product.save().then(result => {
             console.log("Updated Product")
             res.redirect('/admin/products');
         });
     })
-        .catch(err => console.log(err))
+    .catch(err => {
+        console.log(err)
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+})
 }
 
 exports.getProducts = (req, res, next)=>{
@@ -135,13 +184,30 @@ exports.getProducts = (req, res, next)=>{
             path: "/admin/products",
             });
         })
-        .catch(err => console.log(err));
+        .catch(err => {
+            console.log(err)
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+        })
 }
 exports.deleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId,userId: req.user._id})    
+    Product.findById(prodId)
+        .then(product => {
+            if(!product){
+                return next(new Error('Product not Found.'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({_id: prodId,userId: req.user._id})    
+        })
         .then(() => {
             res.redirect('/admin/products');           
         })
-        .catch(err => console.log(err))
+        .catch(err => {
+            console.log(err)
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+       })
 }
